+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Efficiency classes for COP, EER, SCOP, SEER (issue #285).
 - Added building laod classes (issue #288).
 - Added __eq__ method for Result and Efficiency classes (issue #288).
+- Added _time_array to building loads (issue #291).
 
 ## Changed